requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-core==1.16.0
azure-storage-blob==12.8.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
pycparser==2.20
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.6
